feat: Use sdk context propagator for evaluation context
