--- packages/nest/src/async-storage.ts
@@ -1,5 +1,0 @@
-import { AsyncLocalStorage } from 'async_hooks';
-import { AsyncContextType } from './open-feature.module';
-
-export type ContextAsyncStorage = AsyncLocalStorage<AsyncContextType>;
-export const SharedAsyncLocalStorage = new AsyncLocalStorage<AsyncContextType>();
